Moved EmpytResultSet to django.core.exceptions
